Remove code for long/long[].
We might add it back later. A backup exists in the tag 'long-backup'.

+++ src/test/resources/noncompile/symbol/FieldDeclaration.java
@@ -4,6 +4,6 @@ class FieldDeclarationTest {
 }
 class A {
     int a;
-    long a; // DUPLICATE_FIELD
+    int a; // DUPLICATE_FIELD
     B b; // UNDECLARED_CLASS
 }
