+++ src/test/resources/noncompile/symbol/FormalParameter.java
@@ -4,7 +4,7 @@ class FormalParameterTest {
 }
 
 class A {
-    public int f(int b, long b /* DUPLICATE_PARAMETER */, long[] c, boolean d) {
+    public int f(int b, int b /* DUPLICATE_PARAMETER */, int[] c, boolean d) {
         return 42;
     }
     public int g(B b) { // UNDECLARED_CLASS
